requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
bcrypt==4.3.0
pyserial==3.5
python-dotenv==1.0.1
reportlab==4.4.3
Flask-WTF==1.2.1
waitress==3.0.0
# pyzkfp (Installed manually or via wheel likely)